Compile without warnings on Linux. Tom Duggan is fixing BuildBot
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@120492 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/coreskeleton.cpp
@@ -32,7 +32,7 @@ std::vector<CalCoreBonePtr> findInvalidParents(
         
         path.insert(p);
         while (p->parentId != -1) {
-            if (p->parentId >= bones.size()) {
+            if (size_t(p->parentId) >= bones.size()) {
                 knownInvalid.insert(knownInvalid.end(), path.begin(), path.end());
                 break;
             }
@@ -104,7 +104,7 @@ CalCoreSkeleton::CalCoreSkeleton(const std::vector<CalCoreBonePtr>& bones)
 }
 
 size_t CalCoreSkeleton::addCoreBone(const CalCoreBonePtr& coreBone) {
-    cal3d::verify(coreBone->parentId == -1 || coreBone->parentId < coreBones.size(), "bones must be added in topological order");
+    cal3d::verify(coreBone->parentId == -1 || size_t(coreBone->parentId) < coreBones.size(), "bones must be added in topological order");
 
     // skeletons can only have one root
     if (!m_coreBones.empty() && coreBone->parentId == -1) {
